move failed merged usersync package to sync
