+++ pkg/services/authn/authnimpl/sync/apikey_last_seen_sync.go
@@ -1,4 +1,4 @@
-package usersync
+package sync
 
 import (
 	"context"
